+++ lib/Braintree/WebhookNotification.php
@@ -30,7 +30,11 @@ class WebhookNotification extends Base
     const CONNECTED_MERCHANT_PAYPAL_STATUS_CHANGED = 'connected_merchant_paypal_status_changed';
     const IDEAL_PAYMENT_COMPLETE = 'ideal_payment_complete';
     const IDEAL_PAYMENT_FAILED = 'ideal_payment_failed';
+    // NEXT_MAJOR_VERSION remove GRANTED_PAYMENT_INSTRUMENT_UPDATE. Kind is not sent by Braintree Gateway.
+    // Kind will either be GRANTOR_UPDATED_GRANTED_PAYMENT_METHOD or RECIPIENT_UPDATED_GRANTED_PAYMENT_METHOD.
     const GRANTED_PAYMENT_INSTRUMENT_UPDATE = 'granted_payment_instrument_update';
+    const GRANTOR_UPDATED_GRANTED_PAYMENT_METHOD = 'grantor_updated_granted_payment_method';
+    const RECIPIENT_UPDATED_GRANTED_PAYMENT_METHOD = 'recipient_updated_granted_payment_method';
     const GRANTED_PAYMENT_METHOD_REVOKED = 'granted_payment_method_revoked';
     const LOCAL_PAYMENT_COMPLETED = "local_payment_completed";